changed coordinates to BigDecimal

+++ Data Collector/DataCollectorCommons/src/main/java/com/netcracker/datacollector/data/model/bean/Point.java
@@ -3,11 +3,13 @@ package com.netcracker.datacollector.data.model.bean;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class Point {
-    private double x;
-    private double y;
+    private BigDecimal x;
+    private BigDecimal y;
 
     @Override
     public String toString() {
